Move some types over, comments
